tidied up classes
