+++ src/dpr/components/_reports/report-filters/view.njk
@@ -1,5 +1,5 @@
 {% from "../../accordion/view.njk" import dprAccordion %}
-{% from "../../interactive-filters/view.njk" import dprInteractiveFilters %}
+{% from "../../_filters/filters-interactive/view.njk" import dprInteractiveFilters %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
 {% macro dprReportFilters(filterOptions) %}
