+++ src/dpr/utils/renderAsyncReport.ts
@@ -4,7 +4,7 @@ import { AsyncReportUtilsParams } from '../types/AsyncReportUtils'
 import { LoadType, ReportType, RequestedReport } from '../types/UserReports'
 import AsyncReportListUtils from '../components/_async/async-report/utils'
 import ReportActionsUtils from '../components/_reports/report-actions/utils'
-import FiltersUtils from '../components/interactive-filters/utils'
+import FiltersUtils from '../components/_filters/filters-interactive/utils'
 import { Template } from '../types/Templates'
 import ReportQuery from '../types/ReportQuery'
 import CollatedSummaryBuilder from './CollatedSummaryBuilder/CollatedSummaryBuilder'
